pyproject.toml
[project]
name = "project-semantic-layer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "chromadb>=1.5.5",
    "duckdb>=1.4.4",
    "fastapi>=0.128.8",
    "jupyter>=1.1.1",
    "numpy>=2.0.2",
    "ollama>=0.6.1",
    "pandas>=2.3.3",
    "pyyaml>=6.0.3",
    "uvicorn>=0.39.0",
]
